Add files via upload
For the single_pure.py that one is used to send induvial claims to a Discord via webhook. For the multi_pure.py takes a usernames.txt which will send all those claims in that txt file to a Discord via webhook. 

Their is 2 webhook links, used for if you like to post your claims in 2 servers. (Added on both single_pure.py & multi_pure.py)